requirements.txt
langchain-core
langgraph
python-dotenv
youtube-transcript-api
langchain-openai
streamlit
yt-dlp
